Disable Player
The player will now be disabled for a time after a button is pressed.
Specify how long (in seconds) on the button object.

+++ ThrowawayProject/Assets/_Scripts/MyLight.cs
@@ -4,10 +4,10 @@ using System.Collections;
 public interface MyLight {
 	bool GetIsOn ();
 	float GetRange();
-	void FadeOut(int time);
-	void FadeIn(int time);
-	void Shrink (int time);
-	void UnShrink (int time);
+	void FadeOut(float time);
+	void FadeIn(float time);
+	void Shrink (float time);
+	void UnShrink (float time);
 	void FlickerOn (float time);
 	void FlickerOff (float time);
 	void FlickerOn ();
